requirements.txt
amqp==2.3.2
astroid==1.6.5
autopep8==1.3.5
Babel==2.5.3
beautifulsoup4==4.6.0
billiard==3.5.0.4
celery==4.2.1
certifi==2018.1.18
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
coverage==4.4.2
Django==2.0.2
django-admin-rangefilter==0.3.1
django-celery-results==1.0.4
django-ckeditor==5.4.0
django-cool-pagination==0.3.0
django-cors-headers==3.0.2
django-datetimepicker==3.14
django-dbbackup==3.2.0
django-debug-toolbar==1.9.1
django-grappelli==2.11.1
django-htmlmin==0.10.0
django-js-asset==1.0.0
django-json-widget==0.2.0
django-jsonfield==1.3.1
django-modeltranslation==0.13.3
django-mptt==0.9.0
django-nose==1.4.5
django-phonenumber-field==2.0.0
django-robots==3.0
django-social-share==1.3.0
django-widget-tweaks==1.4.1
djangorestframework==3.9.4
drf-yasg==1.15.0
et-xmlfile==1.0.1
facebookads==2.11.4
html5lib==1.0.1
idna==2.6
inflection==0.3.1
isort==4.3.4
itypes==1.1.0
jdcal==1.4
Jinja2==2.10.1
kombu==4.2.1
lazy-object-proxy==1.3.1
lxml==4.1.1
MarkupSafe==1.1.1
mccabe==0.6.1
MonthDelta==0.9.1
mysqlclient==1.3.12
nose==1.3.7
olefile==0.44
openpyxl==2.5.5
phonenumberslite==8.9.0
Pillow==5.0.0
psycopg2==2.7.4
pycodestyle==2.4.0
pylint==1.9.2
pytz==2018.3
requests==2.18.4
ruamel.yaml==0.15.97
six==1.11.0
sorl-thumbnail==12.4.1
sqlparse==0.2.4
uritemplate==3.0.0
urllib3==1.22
vine==1.1.4
webencodings==0.5.1
wrapt==1.10.11
XlsxWriter==1.0.2
